+++ install/upgrade_ajax.php
@@ -685,7 +685,7 @@ if (isset($post_type)) {
                     $filename,
                     $filename.'.'.date(
                         "Y_m_d",
-                        mktime(0, 0, 0, date('m'), date('d'), date('y'))
+                        mktime(0, 0, 0, (int)date('m'), (int)date('d'), (int)date('y'))
                     )
                 )) {
                     echo 'document.getElementById("res_step5").innerHTML = '.
@@ -802,7 +802,7 @@ if (file_exists(\"".$skFile."\")) {
                 $csrfp_file_sample = "../includes/libraries/csrfp/libs/csrfp.config.sample.php";
                 $csrfp_file = "../includes/libraries/csrfp/libs/csrfp.config.php";
                 if (file_exists($csrfp_file) === true) {
-                    if (!copy($filename, $filename.'.'.date("Y_m_d", mktime(0, 0, 0, date('m'), date('d'), date('y'))))) {
+                    if (!copy($filename, $filename.'.'.date("Y_m_d", mktime(0, 0, 0, (int)date('m'), (int)date('d'), (int)date('y'))))) {
                         echo '[{"error" : "csrfp.config.php file already exists and cannot be renamed. Please do it by yourself and click on button Launch.", "result":"", "index" : "'.$post_index.'", "multiple" : "'.$post_multiple.'"}]';
                         break;
                     } else {
