+++ install/upgrade_run_2.1.27.php
@@ -547,7 +547,7 @@ if (isset($session_tp_defuse_installed) === false || $session_tp_defuse_installe
 
     copy(
         SECUREPATH."/teampass-seckey.txt",
-        SECUREPATH."/teampass-seckey.txt".'.'.date("Y_m_d", mktime(0, 0, 0, date('m'), date('d'), date('y'))).".".time()
+        SECUREPATH."/teampass-seckey.txt".'.'.date("Y_m_d", mktime(0, 0, 0, (int)date('m'), (int)date('d'), (int)date('y'))).".".time()
     );
     $superGlobal->put("tp_defuse_new_key", true, "SESSION");
     $new_salt = defuse_generate_key();
@@ -560,7 +560,7 @@ if (isset($session_tp_defuse_installed) === false || $session_tp_defuse_installe
     // update sk.php file
     copy(
         $session_sk_file,
-        $session_sk_file.'.'.date("Y_m_d", mktime(0, 0, 0, date('m'), date('d'), date('y'))).".".time()
+        $session_sk_file.'.'.date("Y_m_d", mktime(0, 0, 0, (int)date('m'), (int)date('d'), (int)date('y'))).".".time()
     );
     $data = file($session_sk_file); // reads an array of lines
     function replace_a_line($data)
@@ -1102,7 +1102,7 @@ while ($row_field = mysqli_fetch_assoc($result)) {
 */
 $tp_config_file = "../includes/config/tp.config.php";
 if (file_exists($tp_config_file)) {
-    if (!copy($tp_config_file, $tp_config_file.'.'.date("Y_m_d", mktime(0, 0, 0, date('m'), date('d'), date('y'))))) {
+    if (!copy($tp_config_file, $tp_config_file.'.'.date("Y_m_d", mktime(0, 0, 0, (int)date('m'), (int)date('d'), (int)date('y'))))) {
         echo '[{"error" : "includes/config/tp.config.php file already exists and cannot be renamed. Please do it by yourself and click on button Launch.", "result":"", "index" : "'.$post_index.'", "multiple" : "'.$post_multiple.'"}]';
         return false;
     } else {
